requirements.txt
fastapi
uvicorn
geopandas
groq
python-dotenv
pandas
torch
scikit-learn
joblib
shapely
pydantic